+++ liger-kernels/torch-ext/liger_kernels/__init__.py
@@ -1,12 +1,8 @@
-from . import ext, layers
-from .ext import LigerForCausalLMLossTransformers
-from .layers import CrossEntropyOutput, LigerForCausalLMLoss
+from . import layers
+from .layers import LigerForCausalLMLoss
 
 
 __all__ = [
-    "ext",
     "layers",
     "LigerForCausalLMLoss",
-    "LigerForCausalLMLossTransformers",
-    "CrossEntropyOutput",
 ]
